separação das interfaces em seu próprio arquivo
