+++ src/ex_4/types.ts
@@ -1,0 +1,10 @@
+export interface Person {
+  id: number;
+  name: string;
+  bio: string;
+}
+
+export interface Update {
+  name?: string;
+  bio?: string;
+}
